+++ Wrappers/Python/ccpi/framework/__init__.py
@@ -34,4 +34,3 @@ from .framework import DataProcessor
 from .framework import AX, PixelByPixelDataProcessor, CastDataContainer
 from .BlockDataContainer import BlockDataContainer
 from .BlockGeometry import BlockGeometry
-from .TestData import TestData
