+++ Wrappers/Python/ccpi/utilities/__init__.py
@@ -16,7 +16,3 @@
 #   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
-#default nThreads
-
-import multiprocessing
-NUM_THREADS = max(int(multiprocessing.cpu_count()/2),1)
