+++ Wrappers/Python/test/test_TestData.py
@@ -18,7 +18,7 @@
 from __future__ import division
 
 import numpy
-from ccpi.framework import TestData
+from ccpi.utilities.TestData import TestData
 import os, sys
 sys.path.append(os.path.dirname(os.path.abspath(__file__)))
 from testclass import CCPiTestClass
